docs: Updated docs around intelligent tariff sensors

+++ custom_components/octopus_energy/api_client.py
@@ -693,7 +693,7 @@ class OctopusEnergyApiClient:
             "planned": list(map(lambda ev: {
                 "start": as_utc(parse_datetime(ev["startDt"])),
                 "end": as_utc(parse_datetime(ev["endDt"])),
-                "charge_in_kwh": int(ev["delta"]) if "delta" in ev and ev["delta"] is not None else None,
+                "charge_in_kwh": float(ev["delta"]) if "delta" in ev and ev["delta"] is not None else None,
                 "source": ev["meta"]["source"] if "meta" in ev and "source" in ev["meta"] else None,
               }, response_body["data"]["plannedDispatches"]
               if "plannedDispatches" in response_body["data"] and response_body["data"]["plannedDispatches"] is not None
@@ -702,7 +702,7 @@ class OctopusEnergyApiClient:
             "completed": list(map(lambda ev: {
                 "start": as_utc(parse_datetime(ev["startDt"])),
                 "end": as_utc(parse_datetime(ev["endDt"])),
-                "charge_in_kwh": int(ev["delta"]) if "delta" in ev and ev["delta"] is not None else None,
+                "charge_in_kwh": float(ev["delta"]) if "delta" in ev and ev["delta"] is not None else None,
                 "source": ev["meta"]["source"] if "meta" in ev and "source" in ev["meta"] else None,
               }, response_body["data"]["completedDispatches"]
               if "completedDispatches" in response_body["data"] and response_body["data"]["completedDispatches"] is not None
